refactor: 게시물 CRUD 코드 리팩토링

+++ board/views.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from board.models import Post
-from board.serializers import PostSimpleSerializer, PostCreateSerializer, PostDetailSerializer, PostPasswordSerializer
+from board.serializers import PostSimpleSerializer, PostCreateSerializer, PostDetailSerializer
 
 
 # Create your views here.
@@ -27,44 +27,37 @@ class PostsView(APIView):
         return Response(serializer_class.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-
 class PostView(APIView):
+    def check_password(self, posting, data):
+        return 'password' in data and posting.password == data.get('password')
+
+    def update_post(self, request, pk, allow_partial=False):
+        posting = get_object_or_404(Post, pk=pk)
+        serializer_class = PostDetailSerializer(posting, data=request.data, partial=allow_partial)
+        if not serializer_class.is_valid():
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        if self.check_password(posting, serializer_class.validated_data):
+            serializer_class.save()
+            return Response(serializer_class.data, status=status.HTTP_201_CREATED)
+        return Response("비밀번호 불일치", status=status.HTTP_403_FORBIDDEN)
+
     def get(self, request, pk):
         posting = get_object_or_404(Post, pk=pk)
         serializer_class = PostDetailSerializer(posting)
         return Response(serializer_class.data)
 
     def post(self, request, pk):
-        posting = get_object_or_404(Post, pk=pk)
-        serializer_class = PostPasswordSerializer(posting, data=request.data, partial=True)
-        if serializer_class.is_valid():
-            return HttpResponseRedirect(reverse("board:post", args=(pk, )))
-        return Response("incorrect password", status=status.HTTP_403_FORBIDDEN)
-
-
-class PostUpdateView(APIView):
-    # def get(self, request, pk):
-    #     posting = get_object_or_404(Post, pk=pk)
-    #     serializer_class = PostDetailSerializer(posting)
-    #     return Response(serializer_class.data)
+        pass
 
     def put(self, request, pk):
-        posting = get_object_or_404(Post, pk=pk)
-        serializer_class = PostDetailSerializer(posting, data=request.data)
-        if serializer_class.is_valid():
-            serializer_class.save()
-            return Response(serializer_class.data, status=status.HTTP_201_CREATED)
-        return Response(serializer_class.errors, status=status.HTTP_400_BAD_REQUEST)
+        return self.update_post(request, pk)
 
     def patch(self, request, pk):
-        posting = get_object_or_404(Post, id=pk)
-        serializer_class = PostDetailSerializer(posting, data=request.data, partial=True)
-        if serializer_class.is_valid():
-            serializer_class.save()
-            return Response(serializer_class.data, status=status.HTTP_201_CREATED)
-        return Response(serializer_class.errors, status=status.HTTP_400_BAD_REQUEST)
+        return self.update_post(request, pk, allow_partial=True)
 
     def delete(self, request, pk):
         posting = get_object_or_404(Post, pk=pk)
-        posting.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if self.check_password(posting, request.GET):
+            posting.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response("비밀번호 불일치", status=status.HTTP_403_FORBIDDEN)
